--- src/main/java/CorruptedFileException.java
@@ -1,5 +1,0 @@
-public class CorruptedFileException extends StorageException {
-    public CorruptedFileException() {
-        super("File Corrupted. FAILURE!");
-    }
-}
