--- src/main/java/CreateFileException.java
@@ -1,5 +1,0 @@
-public class CreateFileException extends StorageException {
-    public CreateFileException() {
-        super("Create file FAILURE!");
-    }
-}
