+++ src/main/java/Duke/Duke.java
@@ -1,0 +1,50 @@
+package Duke;
+
+import Command.Command;
+import DukeException.DukeException;
+import Parser.Parser;
+import Storage.Storage;
+import Task.TaskList;
+import Ui.Ui;
+
+public class Duke {
+
+    private final Storage storage;
+    private TaskList taskList;
+    private final Ui ui;
+
+    public Duke() {
+        this.storage = new Storage("./data/taskList.txt");
+        this.taskList = new TaskList();
+        this.ui = new Ui();
+        try {
+            this.taskList = new TaskList(this.storage.load());
+        } catch (DukeException e) {
+            this.storage.create();
+            ui.showError(e.getMessage());
+            this.taskList = new TaskList();
+        }
+    }
+
+    public void run() {
+        this.ui.sayHi();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                ui.showLine();
+                Command c = Parser.parse(fullCommand);
+                c.execute(this.taskList, this.ui, this.storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e.getMessage());
+            } finally {
+                ui.showLine();
+            }
+        }
+    }
+
+    public static void main(String[] args) {
+        new Duke().run();
+    }
+}
