+++ src/main/java/DukeException/CommandException/EmptyCommandException.java
@@ -1,0 +1,7 @@
+package DukeException.CommandException;
+
+public class EmptyCommandException extends UnknownCommandException {
+    public EmptyCommandException() {
+        super();
+    }
+}
