+++ src/main/java/DukeException/CommandException/UnknownCommandException.java
@@ -1,0 +1,9 @@
+package DukeException.CommandException;
+
+import DukeException.DukeException;
+
+public class UnknownCommandException extends DukeException {
+    public UnknownCommandException() {
+        super("HUH? What you say?", null);
+    }
+}
