+++ src/main/java/DukeException/DukeException.java
@@ -1,0 +1,7 @@
+package DukeException;
+
+public class DukeException extends RuntimeException {
+    public DukeException(String errorMessage, Throwable err) {
+        super(errorMessage, err);
+    }
+}
