+++ src/main/java/DukeException/InvalidDateException.java
@@ -1,0 +1,7 @@
+package DukeException;
+
+public class InvalidDateException extends DukeException {
+    public InvalidDateException() {
+        super("The date wrong format. FAILURE!", null);
+    }
+}
