+++ src/main/java/DukeException/NotFoundException.java
@@ -1,0 +1,7 @@
+package DukeException;
+
+public class NotFoundException extends DukeException {
+    public NotFoundException(String source, String reason, Throwable err) {
+        super(String.format("Haiya nothing here in the %s lah. \n%s", source, reason), err);
+    }
+}
