+++ src/main/java/DukeException/StorageException/CorruptedFileException.java
@@ -1,0 +1,7 @@
+package DukeException.StorageException;
+
+public class CorruptedFileException extends StorageException {
+    public CorruptedFileException() {
+        super("File Corrupted. FAILURE!");
+    }
+}
