+++ src/main/java/DukeException/StorageException/CreateFileException.java
@@ -1,0 +1,7 @@
+package DukeException.StorageException;
+
+public class CreateFileException extends StorageException {
+    public CreateFileException() {
+        super("Create file FAILURE!");
+    }
+}
