+++ src/main/java/DukeException/StorageException/LoadFileException.java
@@ -1,0 +1,7 @@
+package DukeException.StorageException;
+
+public class LoadFileException extends StorageException {
+    public LoadFileException() {
+        super("Load file FAILURE!");
+    }
+}
