+++ src/main/java/DukeException/StorageException/SaveFileException.java
@@ -1,0 +1,7 @@
+package DukeException.StorageException;
+
+public class SaveFileException extends StorageException {
+    public SaveFileException() {
+        super("Save file FAILURE!");
+    }
+}
