+++ src/main/java/DukeException/StorageException/StorageException.java
@@ -1,0 +1,9 @@
+package DukeException.StorageException;
+
+import DukeException.DukeException;
+
+public class StorageException extends DukeException {
+    public StorageException(String message) {
+        super(message, null);
+    }
+}
