--- src/main/java/EmptyCommandException.java
@@ -1,5 +1,0 @@
-public class EmptyCommandException extends UnknownCommandException {
-    public EmptyCommandException() {
-        super();
-    }
-}
