--- src/main/java/InvalidDateException.java
@@ -1,5 +1,0 @@
-public class InvalidDateException extends DukeException {
-    public InvalidDateException() {
-        super("The date wrong format. FAILURE!", null);
-    }
-}
