--- src/main/java/ListCommand.java
@@ -1,10 +1,0 @@
-public class ListCommand extends Command {
-    public ListCommand() {
-
-    }
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        taskList.listItems();
-    }
-}
