--- src/main/java/LoadFileException.java
@@ -1,5 +1,0 @@
-public class LoadFileException extends StorageException {
-    public LoadFileException() {
-        super("Load file FAILURE!");
-    }
-}
