--- src/main/java/NotFoundException.java
@@ -1,5 +1,0 @@
-public class NotFoundException extends DukeException {
-    public NotFoundException(String source, String reason, Throwable err) {
-        super(String.format("Haiya nothing here in the %s lah. \n%s", source, reason), err);
-    }
-}
