--- src/main/java/SaveFileException.java
@@ -1,5 +1,0 @@
-public class SaveFileException extends StorageException {
-    public SaveFileException() {
-        super("Save file FAILURE!");
-    }
-}
