+++ src/main/java/Storage/Storage.java
@@ -1,0 +1,68 @@
+package Storage;
+
+import DukeException.StorageException.CorruptedFileException;
+import DukeException.StorageException.CreateFileException;
+import DukeException.StorageException.LoadFileException;
+import DukeException.StorageException.SaveFileException;
+import Struct.Triple;
+import Task.TaskList;
+
+import java.io.File;
+import java.io.FileWriter;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Scanner;
+
+public class Storage {
+    private final String filePath;
+
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    public void create() {
+        try {
+            File folder = new File("./data");
+            boolean folderCreationSuccessful = folder.mkdir();
+            File file = new File(this.filePath);
+            boolean isFileCreated = file.createNewFile();
+        } catch (IOException e) {
+            throw new CreateFileException();
+        }
+    }
+
+    public void save(TaskList taskList) {
+        try {
+            FileWriter writer = new FileWriter(this.filePath);
+            writer.write(taskList.tasksStorageString("\n"));
+            writer.close();
+            System.out.println("OK saved for you.");
+        } catch (IOException e) {
+            throw new SaveFileException();
+        }
+    }
+
+    public ArrayList<Triple<Character, Boolean, String>> load() {
+        try {
+            File file = new File(this.filePath);
+            Scanner scanner = new Scanner(file);
+            ArrayList<Triple<Character, Boolean, String>> triples = new ArrayList<>();
+            while (scanner.hasNextLine()) {
+                String line = scanner.nextLine();
+                char taskType = line.charAt(0);
+                boolean isMarked = line.charAt(2) == '1';
+                String command = line.substring(4);
+                triples.add(new Triple<>(taskType, isMarked, command));
+            }
+            return triples;
+        } catch (FileNotFoundException e) {
+            throw new LoadFileException();
+        } catch (StringIndexOutOfBoundsException e) {
+            File file = new File(this.filePath);
+            boolean isDeleted = file.delete();
+            this.create();
+            throw new CorruptedFileException();
+        }
+    }
+}
