--- src/main/java/StorageException.java
@@ -1,5 +1,0 @@
-public class StorageException extends DukeException {
-    public StorageException(String message) {
-        super(message, null);
-    }
-}
