+++ src/main/java/Struct/Triple.java
@@ -1,0 +1,25 @@
+package Struct;
+
+public final class Triple<A, B, C> {
+    private final A a;
+    private final B b;
+    private final C c;
+
+    public Triple(A a, B b, C c) {
+        this.a = a;
+        this.b = b;
+        this.c = c;
+    }
+
+    public A getFirst() {
+        return this.a;
+    }
+
+    public B getSecond() {
+        return this.b;
+    }
+
+    public C getThird() {
+        return this.c;
+    }
+}
