+++ src/main/java/Task/Todo.java
@@ -1,0 +1,29 @@
+package Task;
+
+import Parser.Parser;
+
+public class Todo extends Task {
+
+    public Todo(String content) {
+        super(content);
+    }
+
+    public static Todo create(String content) {
+        /**
+         * @param content what to place in this task.
+         * @returns the output Task.Todo object.
+         */
+        Parser.handleEmptyField(content, "content", "Task.Todo Creation");
+        return new Todo(content);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toStorageString() {
+        return String.format("T|%d|%s", this.isMarked() ? 1 : 0, this.getContent());
+    }
+}
