--- src/main/java/TaskType.java
@@ -1,5 +1,0 @@
-public enum TaskType {
-    DEADLINE,
-    EVENT,
-    TODO;
-}
