--- src/main/java/UnknownCommandException.java
@@ -1,5 +1,0 @@
-public class UnknownCommandException extends DukeException {
-    public UnknownCommandException() {
-        super("HUH? What you say?", null);
-    }
-}
